# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: ExpenseVault
  description: Receipt-based expense tracking with OCR automation and categorization
  version: 1.0.0
  contact:
    name: Bernhard M.
    email: bernhard@bermeb.dev
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://expensevault.bermeb.dev/v1
    description: Production server

# For now we are not using OAuth2 or any authentification and leave it for later
#security:
#  - oauth2: [read, write]
paths:
...
  /receipts:
    get:
      tags: [ Receipts ]
      summary: Get all receipts
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            description: Filter by category name
        - name: dateFrom
          in: query
          description: Start date for filtering receipts
          required: false
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: End date for filtering receipts
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReceiptResponse'
        400:
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [ Receipts ]
      summary: Upload and process receipt
      description: Upload a receipt image file for processing. Only the extracted structured dats is stored permanently.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ file ]
              properties:
                file:
                  type: string
                  format: binary
                  description: Receipt image file (JPG; PNG, PDF max 10MB)
                category:
                  type: string
                  description: Optional category override
                description:
                  type: string
                  description: Optional manual description
      responses:
        201:
          description: Receipt processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        400:
          description: Invalid file or processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_file:
                  summary: Invalid file type
                  value:
                    error: "INVALID_FILE"
                    message: "Unsupported file type: GIF. Supported types are: PNG, JPG and PDF"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts"
                category_not_found:
                  summary: Category not found
                  value:
                    error: "CATEGORY_NOT_FOUND"
                    message: "Category not found: InvalidCategory"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts"
        413:
          description: File size exceeds maximum limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                summary: File size exceeded
                value:
                  error: "FILE_SIZE_EXCEEDED"
                  message: "File size exceeds the maximum limit of 10MB"
                  timestamp: "2025-08-22T10:30:00Z"
                  path: "/api/v1/receipts"
        500:
          description: OCR processing failed or internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ocr_failed:
                  summary: OCR processing failed
                  value:
                    error: "OCR_PROCESSING_FAILED"
                    message: "Google Vision API error: Quota exceeded"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts"
                processing_failed:
                  summary: Receipt processing failed
                  value:
                    error: "RECEIPT_PROCESSING_FAILED"
                    message: "Receipt processing failed"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts"
  /receipts/{receiptId}:
    get:
      tags: [ Receipts ]
      summary: Get receipt by ID
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        400:
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "RECEIPT_NOT_FOUND"
                message: "Receipt with ID 123e4567-e89b-12d3-a456-426614174000 not found"
                timestamp: "2025-08-22T10:30:00Z"
                path: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ Receipts ]
      summary: Update receipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptUpdateRequest'
      responses:
        200:
          description: Receipt updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        400:
          description: Invalid input data or UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_failed:
                  summary: Validation failed
                  value:
                    error: "VALIDATION_FAILED"
                    message: "Validation failed: {amount=must be greater than 0}"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000"
                category_not_found:
                  summary: Category not found
                  value:
                    error: "CATEGORY_NOT_FOUND"
                    message: "Category with ID 456e7890-e89b-12d3-a456-426614174000 not found"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000"
        404:
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ Receipts ]
      summary: Delete receipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Receipt deleted
        400:
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /receipts/{receiptId}/reprocess:
    post:
      tags: [ Receipts ]
      summary: Reprocess receipt with new image
      description: Upload a new image to reprocess OCR data for existing receipt
      parameters:
        - name: receiptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ file ]
              properties:
                file:
                  type: string
                  format: binary
                  description: New receipt image file
      responses:
        200:
          description: Receipt reprocessed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        400:
          description: Invalid file format or UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_file:
                  summary: Invalid file type
                  value:
                    error: "INVALID_FILE"
                    message: "Only JPG, PNG, and PDF files are allowed"
                    timestamp: "2025-08-22T10:30:00Z"
                    path: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000/reprocess"
        404:
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        413:
          description: File size exceeds maximum limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                summary: File size exceeded
                value:
                  error: "FILE_SIZE_EXCEEDED"
                  message: "File size exceeds the maximum limit of 10MB"
                  timestamp: "2025-08-22T10:30:00Z"
                  path: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000/reprocess"
        500:
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ocr_failed:
                  summary: OCR processing failed
                  value:
                    error: "OCR_PROCESSING_FAILED"
                    message: "Google Vision API error: Quota exceeded"
                    timestamp: "2025-08-22T10:30:00Z"

  # Category section
...
components:
...
  schemas:
    ReceiptResponse:
      type: object
      required:
        - id
        - amount
        - currency
        - date
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
          description: Total amount of the receipt
          example: 24.99
        currency:
          type: string
          default: EUR
        merchantName:
          type: string
          example: "XY Supermarket"
        description:
          type: string
          example: "Weekly groceries"
        date:
          type: string
          format: date
          example: "2025-08-22"
        category:
          $ref: '#/components/schemas/CategoryResponse'
        ocrData:
          $ref: '#/components/schemas/OcrDataResponse'
        createdAt:
          type: string
          format: date-time
          example: "2025-08-22T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-22T11:00:00Z"
    ReceiptUpdateRequest:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0
        merchantName:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        date:
          type: string
          format: date
        categoryId:
          type: string
          format: uuid
    OcrDataResponse:
      type: object
      properties:
        rawText:
          type: string
          description: Complete extracted text
          example: "XY Supermarket\nReceiptNr: 1234\nDate: 22.08.2025\nTotal: 24,99 EUR"
        confidence:
          type: number
          format: float
          description: OCR confidence score (0-1)
          minimum: 0
          maximum: 1
          example: 0.96
        processedAt:
          type: string
          format: date-time
          example: "2025-08-22T10:00:05Z"
        extractedFields:
          type: object
          description: Structured data extracted from text
          properties:
            detectedAmount:
              type: number
              format: decimal
            detectedDate:
              type: string
              format: date
            detectedMerchant:
              type: string
            detectedItems:
              type: array
              items:
                type: string
...
    CategoryResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Groceries"
        color:
          type: string
          description: Hex color code
          example: "#FF5733"
        icon:
          type: string
          description: Icon identifier
          example: "shopping_cart"
        receiptCount:
          type: integer
          description: Number of receipts in this category
          example: 25
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - path
      properties:
        error:
          type: string
          example: "RECEIPT_NOT_FOUND"
        message:
          type: string
          example: "Receipt with ID 123e4567-e89b-12d3-a456-426614174000 not found"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-22T10:30:00Z"
        path:
          type: string
          example: "/api/v1/receipts/123e4567-e89b-12d3-a456-426614174000"